Energy ui, refactoring
Wrote the logic of energy ui, fixed bugs, refactored code

+++ Arcanoid Maket/Assets/Scripts/MyLibrary/EnergySystem/Data/Config/EnergyConfig.cs
@@ -13,12 +13,12 @@ namespace MyLibrary.EnergySystem.Data.Config
         private int _energyPerTime = 1;
 
         [SerializeField]
-        private int _restoringTimeInSeconds = 60;
+        private float _restoringTimeInSeconds = 60;
 
         [SerializeField]
         private EnergyValuesContainer _energyValuesContainer;
         
-        public int RestoringTimeInSeconds => _restoringTimeInSeconds;
+        public float RestoringTimeInSeconds => _restoringTimeInSeconds;
         public int EnergyPerTime => _energyPerTime;
         public int MaxEnergy => _maxEnergy;
         public EnergyValuesContainer EnergyValuesContainer => _energyValuesContainer;
